bilbo contribution!! Mummi contribution!! veriverse contribution!! Catherina Cortega, Benjamin Kuhl,  Pawel Kulesza, chubz, diagram, hudwan, intelijens, bigboom, dopedopedop, leuus, mahmudul0006, shanshanren, kaeris,  kissloryshy, kord and kranalex contribution!! Twtich contribution!! Connavardev, Gaseline, Kurroqt, Landavaranda and Odielag contribution!!
Thank you Carlos!!

Thank you Mummi!!

Thank you Carol!!

from bilbo Ubuntu!!
i7 6700 32GB RAM

+++ appseed/aura/multithreading/service/service_plain_service.cpp
@@ -20,7 +20,7 @@ void plain_service::Stop(uint32_t ui)
 {
    System.m_bDoNotExitIfNoApplications = false;
    m_bStopping = true;
-   m_stopped.wait(millis((long long)m_dwStopTimeout));
+   m_stopped.wait(millis(m_dwStopTimeout));
 }
 
 void plain_service::CallServiceThread()
